requirements.txt
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.6
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.12.0
identify==2.5.24
importlib-metadata==6.6.0
Markdown==3.4.3
nodeenv==1.8.0
platformdirs==3.5.1
pre-commit==2.21.0
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
virtualenv==20.23.0
zipp==3.15.0
